requirements.txt
# Parking Availability Prediction AI
# Python 3.10+

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Optional: HTTP client for real data
httpx>=0.25.0
PyYAML>=6.0

# Dev / optional
pytest>=7.4.0
